+++ flash/text/classification/input.py
@@ -16,8 +16,8 @@ from typing import Any, Dict, List, Optional, Union
 
 import pandas as pd
 
-from flash.core.data.io.classification_input import ClassificationInput, ClassificationState
-from flash.core.data.io.input import DataKeys
+from flash.core.data.io.classification_input import ClassificationInputMixin, ClassificationState
+from flash.core.data.io.input import DataKeys, Input
 from flash.core.data.utilities.classification import TargetMode
 from flash.core.data.utilities.paths import PATH_TYPE
 from flash.core.integrations.transformers.states import TransformersBackboneState
@@ -29,7 +29,7 @@ else:
     Dataset = object
 
 
-class TextClassificationInput(ClassificationInput):
+class TextClassificationInput(Input, ClassificationInputMixin):
     @staticmethod
     def _resolve_target(target_keys: Union[str, List[str]], element: Dict[str, Any]) -> Dict[str, Any]:
         if not isinstance(target_keys, List):